+++ packages/react/src/composite/Card/src/card.stories.tsx
@@ -64,7 +64,7 @@ WithHeader.args = {
         <Box
           borderStyle="dashed"
           padding="2"
-          borderWidth="1px"
+          borderWidth="1"
           borderColor="neutral-interactive"
         >
           <Text textAlign="center">Replace me with your content</Text>
@@ -83,7 +83,7 @@ WithFooterAndHeader.args = {
         <Box
           borderStyle="dashed"
           padding="2"
-          borderWidth="1px"
+          borderWidth="1"
           borderColor="neutral-interactive"
         >
           <Text textAlign="center">Replace me with your content</Text>